fix: variables index when parser equations
Before, the index was define in the order that the variables apper
in the equations, but we have the list of variables between the
JSON, so that order was in conflict with the first indexing.

Now, we save the variables names of each equation, and when create
the equation object find the corresponding index to that name from
the variables JSON index.